Executing migrate if no assets found

+++ src/migrate/index.js
@@ -3,15 +3,39 @@ const Market = require('../models/Market')
 const assets = require('./data/assets.json')
 const markets = require('./data/markets.json')
 
-module.exports.run = async () => {
-  console.log('[DEVING] Running migrate...')
-  await Asset.deleteMany({})
-  const newAssets = await Asset.insertMany(assets, { ordered: false })
-  console.log(`${newAssets.length} assets have been set`)
+const logHeader = '[MIGRATE]'
 
-  await Market.deleteMany({})
-  const newMarkets = await Market.insertMany(markets, { ordered: false })
-  console.log(`${newMarkets.length} markets have been set`)
+module.exports.run = async (options = {}) => {
+  console.log(`${logHeader} Running migrate with options:`, options)
 
-  // process.exit()
+  if (await hasData()) {
+    console.log(`${logHeader} Data is already seeded.`)
+  } else {
+    console.log(`${logHeader} Seeding data...`)
+    await Asset.deleteMany({})
+    const newAssets = await Asset.insertMany(assets, { ordered: false })
+    console.log(`${newAssets.length} assets have been set`)
+
+    await Market.deleteMany({})
+    const newMarkets = await Market.insertMany(markets, { ordered: false })
+    console.log(`${newMarkets.length} markets have been set`)
+  }
+}
+
+async function hasData () {
+  let result = true
+
+  console.log(`${logHeader} Checking for existing Assets`)
+  // console.log(`${logHeader}:`, db.runCommand( { listCollections: 1 } ))
+
+  const existingAssets = await Asset.find({}).exec()
+
+  if (!existingAssets || existingAssets.length ===  0) {
+    console.log('No assets found.')
+    result = false
+  } else {
+    console.log('Existing assets:', existingAssets)
+  }
+
+  return result
 }
